requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.8.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-debug-toolbar==4.4.6
kombu==5.4.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
gunicorn
redis
flower
stripe